fix: rename enrichment table migration file name
