+++ src/infra/src/table/migration/m20250923_000001_update_enrichment_table_data_size.rs
