+++ src/infra/src/table/migration/mod.rs
@@ -64,7 +64,7 @@ mod m20250716_000001_create_enrichment_table;
 mod m20250731_000001_create_compactor_manual_jobs;
 mod m20250801_000001_create_system_prompts_table;
 mod m20250822_093713_add_updated_at_for_file_list_table;
-mod m20251023_000001_update_enrichment_table_data_size;
+mod m20250923_000001_update_enrichment_table_data_size;
 
 pub struct Migrator;
 
@@ -121,7 +121,7 @@ impl MigratorTrait for Migrator {
             Box::new(m20250731_000001_create_compactor_manual_jobs::Migration),
             Box::new(m20250801_000001_create_system_prompts_table::Migration),
             Box::new(m20250822_093713_add_updated_at_for_file_list_table::Migration),
-            Box::new(m20251023_000001_update_enrichment_table_data_size::Migration),
+            Box::new(m20250923_000001_update_enrichment_table_data_size::Migration),
         ]
     }
 }
